refactor: variant 모듈 분리
- variant가 product 애그리거트와 분리됨에 따라, variant 모듈로 분리하였음
