+++ src/main/java/kr/tatine/manibogo_oms_v2/order/ui/OrderController.java
@@ -5,8 +5,8 @@ import kr.tatine.manibogo_oms_v2.order.command.domain.model.vo.OrderState;
 import kr.tatine.manibogo_oms_v2.order.command.domain.model.vo.SalesChannel;
 import kr.tatine.manibogo_oms_v2.order.query.dao.OrderDao;
 import kr.tatine.manibogo_oms_v2.order.query.dto.*;
-import kr.tatine.manibogo_oms_v2.product.query.dao.ProductDao;
-import kr.tatine.manibogo_oms_v2.product.query.dto.ProductDto;
+import kr.tatine.manibogo_oms_v2.product.query.ProductDao;
+import kr.tatine.manibogo_oms_v2.product.query.ProductDto;
 import kr.tatine.manibogo_oms_v2.region.query.dao.GroupRegionsService;
 import kr.tatine.manibogo_oms_v2.synchronize.ui.SynchronizeResponse;
 import lombok.RequiredArgsConstructor;
