+++ src/main/java/kr/tatine/manibogo_oms_v2/product/query/ProductDao.java
@@ -1,8 +1,7 @@
-package kr.tatine.manibogo_oms_v2.product.query.dao;
+package kr.tatine.manibogo_oms_v2.product.query;
 
 import kr.tatine.manibogo_oms_v2.product.command.domain.Product;
 import kr.tatine.manibogo_oms_v2.product.command.domain.ProductNumber;
-import kr.tatine.manibogo_oms_v2.product.query.dto.ProductDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
@@ -15,7 +14,7 @@ import java.util.Optional;
 public interface ProductDao extends Repository<Product, ProductNumber> {
 
     @Query("""
-        SELECT new kr.tatine.manibogo_oms_v2.product.query.dto.ProductDto(
+        SELECT new kr.tatine.manibogo_oms_v2.product.query.ProductDto(
             p.number.productNumber,
             p.name,
             p.priority.priority,
@@ -27,21 +26,21 @@ public interface ProductDao extends Repository<Product, ProductNumber> {
     List<ProductDto> findAll();
 
     @Query("""
-        SELECT new kr.tatine.manibogo_oms_v2.product.query.dto.ProductDto(p.number.productNumber, p.name, p.priority.priority, p.isEnabled)
+        SELECT new kr.tatine.manibogo_oms_v2.product.query.ProductDto(p.number.productNumber, p.name, p.priority.priority, p.isEnabled)
         FROM Product p
         ORDER BY p.priority.priority
     """)
     Page<ProductDto> findAll(Pageable pageable);
 
     @Query("""
-        SELECT new kr.tatine.manibogo_oms_v2.product.query.dto.ProductDto(p.number.productNumber, p.name, p.priority.priority, p.isEnabled)
+        SELECT new kr.tatine.manibogo_oms_v2.product.query.ProductDto(p.number.productNumber, p.name, p.priority.priority, p.isEnabled)
         FROM Product p
         WHERE p.isEnabled = true
         ORDER BY p.priority.priority
     """)
     List<ProductDto> findEnabled();
 
-    @Query("SELECT new kr.tatine.manibogo_oms_v2.product.query.dto.ProductDto(" +
+    @Query("SELECT new kr.tatine.manibogo_oms_v2.product.query.ProductDto(" +
             "p.number.productNumber, " +
             "p.name, " +
             "p.priority.priority, " +
