+++ src/main/java/kr/tatine/manibogo_oms_v2/product/ui/ProductController.java
@@ -8,8 +8,8 @@ import kr.tatine.manibogo_oms_v2.common.utils.SelectableRowsFormUtils;
 import kr.tatine.manibogo_oms_v2.product.command.application.EditProductService;
 import kr.tatine.manibogo_oms_v2.product.command.application.ProductNotFoundException;
 import kr.tatine.manibogo_oms_v2.product.command.domain.ProductNameDuplicatedException;
-import kr.tatine.manibogo_oms_v2.product.query.dao.ProductDao;
-import kr.tatine.manibogo_oms_v2.product.query.dto.ProductDto;
+import kr.tatine.manibogo_oms_v2.product.query.ProductDao;
+import kr.tatine.manibogo_oms_v2.product.query.ProductDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
