+++ src/main/java/kr/tatine/manibogo_oms_v2/synchronize/command/application/SyncExternalOrderService.java
@@ -10,9 +10,9 @@ import kr.tatine.manibogo_oms_v2.common.model.PhoneNumber;
 import kr.tatine.manibogo_oms_v2.order.command.application.exception.OrderAlreadyPlacedException;
 import kr.tatine.manibogo_oms_v2.order.command.domain.model.vo.ShippingMethod;
 import kr.tatine.manibogo_oms_v2.synchronize.ui.ExternalItemOrderRequest;
-import kr.tatine.manibogo_oms_v2.product.command.domain.Variant;
-import kr.tatine.manibogo_oms_v2.product.command.domain.VariantId;
-import kr.tatine.manibogo_oms_v2.product.command.domain.VariantRepository;
+import kr.tatine.manibogo_oms_v2.variant.command.domain.Variant;
+import kr.tatine.manibogo_oms_v2.variant.command.domain.VariantId;
+import kr.tatine.manibogo_oms_v2.variant.command.domain.VariantRepository;
 import kr.tatine.manibogo_oms_v2.order.command.domain.model.Order;
 import kr.tatine.manibogo_oms_v2.order.command.domain.model.vo.*;
 import kr.tatine.manibogo_oms_v2.order.command.domain.repository.OrderRepository;
