+++ src/main/java/kr/tatine/manibogo_oms_v2/variant/command/application/VariantDuplicatedException.java
@@ -1,3 +1,3 @@
-package kr.tatine.manibogo_oms_v2.product.command.application;
+package kr.tatine.manibogo_oms_v2.variant.command.application;
 
 public class VariantDuplicatedException extends RuntimeException { }
