+++ src/main/java/kr/tatine/manibogo_oms_v2/variant/query/VariantDao.java
@@ -1,8 +1,7 @@
-package kr.tatine.manibogo_oms_v2.product.query.dao;
+package kr.tatine.manibogo_oms_v2.variant.query;
 
-import kr.tatine.manibogo_oms_v2.product.command.domain.Variant;
-import kr.tatine.manibogo_oms_v2.product.command.domain.VariantId;
-import kr.tatine.manibogo_oms_v2.product.query.dto.VariantDto;
+import kr.tatine.manibogo_oms_v2.variant.command.domain.Variant;
+import kr.tatine.manibogo_oms_v2.variant.command.domain.VariantId;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.Query;
@@ -12,7 +11,7 @@ import java.util.List;
 
 public interface VariantDao extends Repository<Variant, VariantId> {
 
-    @Query("SELECT new kr.tatine.manibogo_oms_v2.product.query.dto.VariantDto(" +
+    @Query("SELECT new kr.tatine.manibogo_oms_v2.variant.query.VariantDto(" +
             "   v.variantId.productNumber.productNumber, " +
             "   v.variantId.option.key, " +
             "   v.variantId.option.value, " +
@@ -21,7 +20,7 @@ public interface VariantDao extends Repository<Variant, VariantId> {
             "WHERE v.variantId.productNumber.productNumber = :productNumber")
     Page<VariantDto> findAllByProductNumber(Pageable pageable, String productNumber);
 
-    @Query("SELECT new kr.tatine.manibogo_oms_v2.product.query.dto.VariantDto(" +
+    @Query("SELECT new kr.tatine.manibogo_oms_v2.variant.query.VariantDto(" +
             "   v.variantId.productNumber.productNumber, " +
             "   v.variantId.option.key, " +
             "   v.variantId.option.value, " +
